Refactored the code structure.
